Complete refactor of css into scss
