+++ client/src/assets/styles_unused/Drawers.css
@@ -1,0 +1,89 @@
+* {
+    box-sizing: border-box;
+}
+
+body {
+    font-size: 13px;
+    line-height: normal;
+    color: #040c13;
+    font-family: Human BBY Web,Arial,Helvetica,sans-serif;
+}
+
+h4 {
+    font-weight: bold;
+}
+
+ul {
+    padding: 0;
+}
+
+.body-copy {
+    font-size: 13px;
+    line-height: normal;
+}
+
+.drawers {
+    width: 100%;
+    border-bottom-color: rgb(197, 203, 213);
+    border-bottom-style: solid;
+    border-bottom-width: 1px;
+}
+
+.card {
+    border-top-color: rgb(197, 203, 213);
+    border-top-style: solid;
+    border-top-width: 1px;
+}
+
+.card-header {
+    display: flex;
+    align-items: center;
+    padding-top: 24px;
+    padding-bottom: 24px;
+    font-size: 25px;
+    font-weight: bold;
+    cursor: pointer;
+}
+
+.drawer-label {
+    flex: 1 0 0;
+}
+
+.drawer-caret {
+    flex: 0 0 25px;
+}
+
+.spacer-row {
+    display: flex;
+}
+
+.spacer-row-3, .spacer-row-9 {
+    position: relative;
+    min-height: 1px;
+    margin-left: -10px;
+}
+
+.spacer-row-9 {
+    flex: 3;
+}
+
+.spacer-row-3 {
+    flex: 1;
+}
+
+.spacer-row-1 {
+    flex: 1;
+}
+
+.r-padding-bottom {
+    padding-bottom: 24px;
+}
+
+.r-margin-bottom {
+    margin-bottom: 24px;
+}
+
+.r-border-bottom {
+    border-bottom: 1px solid #c5cbd5;
+}
+
