+++ client/src/components/App.scss
@@ -1,0 +1,32 @@
+#accordion-drawers-container {
+  * {
+    box-sizing: border-box;
+  }
+  div.container {
+    display: flex;
+    justify-content: center;
+    width: 100vw;
+    max-width: 1600px;
+    min-width: 992px;
+    padding-left: 30px !important;
+    padding-right: 30px !important;
+    margin: 0 auto;
+  }
+
+  @media screen and (min-width: 1200px) {
+    div.container {
+      width: 100%;
+    }
+  }
+
+  @media screen and (min-width: 992px) {
+    div.container {
+      width: 100%;
+    }
+  }
+
+  .icon-credit {
+    font-size: 2px;
+    display: none;
+  }
+}
