+++ client/src/components/Drawers.jsx
@@ -1,13 +1,10 @@
 import React from "react";
-import Card from "react-bootstrap/Card";
-import Collapse from "react-bootstrap/Collapse";
 
 import Overview from './overview/Overview.jsx';
 import Specs from './specs/Specs.jsx';
 import ReviewDrawer from './reviews/ReviewDrawer.jsx';
-import QandA from './qanda/QandA.jsx';
 
-import styles from "./Drawers.css";
+import "./Drawers.scss";
 
 export default class Drawers extends React.Component {
   constructor(props) {
@@ -21,7 +18,7 @@ export default class Drawers extends React.Component {
 
   render() {
     return (
-      <div className="drawers">
+      <div className="drawers" id="drawers-container">
         <Overview productId={this.state.productId} />
         <Specs productId={this.state.productId} />
         <ReviewDrawer productId={this.state.productId} />
