+++ client/src/components/RatingHeader.scss
@@ -1,0 +1,209 @@
+#drawer-header-rating {
+    .drawer-header-rating-container {
+        display: inline-block;
+        padding-left: 10px;
+    }
+    
+    .stars {
+        background-image: url(../assets/bby-stars.png);
+        background-repeat: no-repeat;
+        display: inline-block;
+        position: relative;
+        margin-bottom: 0px;
+    }
+    
+    .drawer-header-review-count {
+        display: inline-block;
+        line-height: 18px;
+        padding-left: 5px;
+        vertical-align: top;
+        position: relative;
+    }
+    
+    .drawer-header-total-reviews {
+        font-weight: 400;
+        font-size: 14px;
+        vertical-align: top;
+    }
+    
+    .stars-small {
+        width: 90px;
+        height: 18px;
+    }
+    
+    /*
+    small row height: 26px
+    small star width: 18px
+    */
+    
+    .stars-small-0-0 {
+        background-position: -318px -262px;
+    }
+    
+    .stars-small-5-0 {
+        background-position: -318px -54px;
+    }
+    
+    .stars-small-4-9 {
+        background-position: -318px -80px;
+    }
+    
+    .stars-small-4-8, .stars-small-4-7 {
+        background-position: -318px -106px;
+    }
+    
+    .stars-small-4-6 {
+        background-position: -318px -132px;
+    }
+    
+    .stars-small-4-5 {
+        background-position: -318px -158px;
+    }
+    
+    .stars-small-4-4 {
+        background-position: -318px -184px;
+    }
+    
+    .stars-small-4-3, .stars-small-4-2 {
+        background-position: -318px -210px;
+    }
+    
+    .stars-small-4-1 {
+        background-position: -318px -236px;
+    }
+    
+    .stars-small-4-0 {
+        background-position: -336px -54px;
+    }
+    
+    .stars-small-3-9 {
+        background-position: -336px -80px;
+    }
+    
+    .stars-small-3-7, .stars-small-3-8 {
+        background-position: -336px -106px;
+    }
+    
+    .stars-small-3-6 {
+        background-position: -336px -132px;
+    }
+    
+    .stars-small-3-5 {
+        background-position: -336px -158px;
+    }
+    
+    .stars-small-3-4 {
+        background-position: -336px -184px;
+    }
+    
+    .stars-small-3-3, .stars-small-3-2 {
+        background-position: -336px -210px;
+    }
+    
+    .stars-small-3-1 {
+        background-position: -336px -236px;
+    }
+    
+    .stars-small-3-0 {
+        background-position: -354px -54px;
+    }
+    
+    .stars-small-2-9 {
+        background-position: -354px -80px;
+    }
+    
+    .stars-small-2-7, .stars-small-2-8 {
+        background-position: -354px -106px;
+    }
+    
+    .stars-small-2-6 {
+        background-position: -354px -132px;
+    }
+    
+    .stars-small-2-5 {
+        background-position: -354px -158px;
+    }
+    
+    .stars-small-2-4 {
+        background-position: -354px -184px;
+    }
+    
+    .stars-small-2-3, .stars-small-2-2 {
+        background-position: -354px -210px;
+    }
+    
+    .stars-small-2-1 {
+        background-position: -354px -236px;
+    }
+    
+    .stars-small-2-0 {
+        background-position: -372px -54px;
+    }
+    
+    .stars-small-1-9 {
+        background-position: -372px -80px;
+    }
+    
+    .stars-small-1-7, .stars-small-1-8 {
+        background-position: -372px -106px;
+    }
+    
+    .stars-small-1-6 {
+        background-position: -372px -132px;
+    }
+    
+    .stars-small-1-5 {
+        background-position: -372px -158px;
+    }
+    
+    .stars-small-1-4 {
+        background-position: -372px -184px;
+    }
+    
+    .stars-small-1-3, .stars-small-1-2 {
+        background-position: -372px -210px;
+    }
+    
+    .stars-small-1-1 {
+        background-position: -372px -236px;
+    }
+    
+    .stars-small-1-0 {
+        background-position: -390px -54px;
+    }
+    
+    .stars-small-0-9 {
+        background-position: -390px -80px;
+    }
+    
+    .stars-small-0-7, .stars-small-0-8 {
+        background-position: -390px -106px;
+    }
+    
+    .stars-small-0-6 {
+        background-position: -390px -132px;
+    }
+    
+    .stars-small-0-5 {
+        background-position: -390px -158px;
+    }
+    
+    .stars-small-0-4 {
+        background-position: -390px -184px;
+    }
+    
+    .stars-small-0-3, .stars-small-0-2 {
+        background-position: -390px -210px;
+    }
+    
+    .stars-small-0-1 {
+        background-position: -390px -236px;
+    }
+}
+
+#drawer-header-rating.drawer-header-rating-container {
+    display: inline-flex;
+    padding-left: 10px;
+    height: 29px;
+    align-items: center;
+}
