+++ client/src/index.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './components/App.jsx';
+import Drawers from './components/Drawers.jsx';
 import StarGraphic from './components/StarGraphic.jsx';
 
 ReactDOM.render(<App />, document.getElementById('accordion-drawers-container'));
